move Java backend-specific passes out of default
