+++ kernel/src/main/java/org/kframework/kompile/Kompile.java
@@ -153,8 +153,6 @@ public class Kompile {
         DefinitionTransformer resolveSemanticCasts =
                 DefinitionTransformer.fromSentenceTransformer(new ResolveSemanticCasts()::resolve, "resolving semantic casts");
         DefinitionTransformer generateSortPredicateSyntax = DefinitionTransformer.from(new GenerateSortPredicateSyntax()::gen, "adding sort predicate productions");
-        DefinitionTransformer convertDataStructureToLookup = DefinitionTransformer.fromSentenceTransformer(func((m, s) -> new ConvertDataStructureToLookup(m, false).convert(s)), "convert data structures to lookups");
-
 
         return def -> func(this::resolveIOStreams)
                 .andThen(resolveStrict)
@@ -168,65 +166,7 @@ public class Kompile {
                 .andThen(func(this::concretizeTransformer))
                 .andThen(func(this::addSemanticsModule))
                 .andThen(func(this::addProgramModule))
-                .andThen(DefinitionTransformer.fromRuleBodyTranformer(RewriteToTop::bubbleRewriteToTop, "bubble out rewrites below cells"))
-                .andThen(convertDataStructureToLookup)
-                //.andThen(DefinitionTransformer.fromRuleBodyTranformer(RewriteToTop::bubbleRewriteOutOfKSeq, "bubble rewrites out of kseq"))
-                .andThen(DefinitionTransformer.fromRuleBodyTranformer(Kompile::ADTKVariableToSortedVariable, "ADT.KVariable to SortedVariable"))
-                .andThen(DefinitionTransformer.fromRuleBodyTranformer(Kompile::convertKSeqToKApply, "kseq to kapply"))
-                .andThen(DefinitionTransformer.fromRuleBodyTranformer(CleanKSeq.self(), "normalize kseq"))
-                .andThen(DefinitionTransformer.fromSentenceTransformer(this::markSingleVariables, "mark single variables"))
-                .andThen(new DefinitionTransformer(new MergeRules(KORE.c())))
                 .apply(def);
-    }
-
-    private static K ADTKVariableToSortedVariable(K ruleBody) {
-        return new TransformKORE() {
-            public K apply(KVariable kvar) {
-                return new SortedADT.SortedKVariable(kvar.name(), kvar.att());
-            }
-        }.apply(ruleBody);
-    }
-
-    private static K convertKSeqToKApply(K ruleBody) {
-        return new TransformKORE() {
-            public K apply(KSequence kseq) {
-                return ((ADT.KSequence) kseq).kApply();
-            }
-        }.apply(ruleBody);
-    }
-
-    private Sentence markSingleVariables(Sentence s) {
-        if (s instanceof Rule) {
-            Rule r = (Rule) s;
-
-            if (!(r.body() instanceof KApply) || !((KApply) r.body()).klabel().name().equals("<T>"))
-                return r;
-
-            Map<KVariable, Integer> varCount = new HashMap<>();
-            VisitKORE markerVisitor = new VisitKORE() {
-                public Void apply(KVariable kvar) {
-                    varCount.put(kvar, varCount.getOrDefault(kvar, 0) + 1);
-                    return null;
-                }
-            };
-            markerVisitor.apply(r.body());
-            markerVisitor.apply(r.requires());
-            markerVisitor.apply(r.ensures());
-
-            TransformKORE markerAdder = new TransformKORE() {
-                public K apply(KVariable kvar) {
-                    if (varCount.get(kvar) == 1) {
-                        return new SortedADT.SortedKVariable("THE_VARIABLE", Att());
-                    } else {
-                        return kvar;
-                    }
-                }
-            };
-
-            return Rule(markerAdder.apply(r.body()), markerAdder.apply(r.requires()), markerAdder.apply(r.ensures()), r.att());
-        } else {
-            return s;
-        }
     }
 
     public Definition resolveIOStreams(Definition d) {
